+++ mne/preprocessing/ica.py
@@ -45,7 +45,7 @@ from ..channels.channels import _contains_ch_type, ContainsMixin
 from ..io.write import start_file, end_file, write_id
 from ..utils import (check_version, logger, check_fname, verbose,
                      _reject_data_segments, check_random_state,
-                     _get_fast_dot, compute_corr,
+                     _get_fast_dot, compute_corr, _get_inst_data,
                      copy_function_doc_to_method_doc)
 from ..fixes import _get_args
 from ..filter import band_pass_filter
@@ -359,6 +359,12 @@ class ICA(ContainsMixin):
                 self._fit_epochs(inst, picks, decim, verbose)
         else:
             raise ValueError('Data input must be of Raw or Epochs type')
+
+        # sort ICA components by explained variance
+        var = _ica_explained_variance(self, inst)
+        var_ord = var.argsort()[::-1]
+        _sort_components(self, var_ord, copy=False)
+
         return self
 
     def _reset(self):
@@ -1622,6 +1628,68 @@ def _find_sources(sources, target, score_func):
               else score_func(sources, 1))
 
     return scores
+
+
+def _ica_explained_variance(ica, inst, normalize=False):
+    """Checks variance accounted for by each component in supplied data.
+
+    Parameters
+    ----------
+    ica : ICA
+        Instance of `mne.preprocessing.ICA`.
+    inst : Raw | Epochs | Evoked
+        Data to explain with ICA. Instance of Raw, Epochs or Evoked.
+    normalize : bool
+        Whether to normalize the variance.
+
+    Returns
+    -------
+    var : array
+        Variance explained by each component.
+    """
+    # check if ica is ICA and whether inst is Raw or Epochs
+    if not isinstance(ica, ICA):
+        raise TypeError('first argument must be an instance of ICA.')
+    if not isinstance(inst, (_BaseRaw, _BaseEpochs, Evoked)):
+        raise TypeError('second argument must an instance of either Raw, '
+                        'Epochs or Evoked.')
+
+    source_data = _get_inst_data(ica.get_sources(inst))
+
+    # if epochs - reshape to channels x timesamples
+    if isinstance(inst, _BaseEpochs):
+        n_epochs, n_chan, n_samp = source_data.shape
+        source_data = source_data.transpose(1, 0, 2).reshape(
+            (n_chan, n_epochs * n_samp))
+
+    n_chan, n_samp = source_data.shape
+    var = np.sum(ica.mixing_matrix_**2, axis=0) * np.sum(
+        source_data**2, axis=1) / (n_chan * n_samp - 1)
+    if normalize:
+        var /= var.sum()
+    return var
+
+
+def _sort_components(ica, order, copy=True):
+    """Change the order of components in ica solution."""
+    assert ica.n_components_ == len(order)
+    if copy:
+        ica = ica.copy()
+
+    # reorder components
+    ica.mixing_matrix_ = ica.mixing_matrix_[:, order]
+    ica.unmixing_matrix_ = ica.unmixing_matrix_[order, :]
+
+    # reorder labels, excludes etc.
+    if isinstance(order, np.ndarray):
+        order = list(order)
+    if ica.exclude:
+        ica.exclude = [order.index(ic) for ic in ica.exclude]
+    if hasattr(ica, 'labels_'):
+        for k in ica.labels_.keys():
+            ica.labels_[k] = [order.index(ic) for ic in ica.labels_[k]]
+
+    return ica
 
 
 def _serialize(dict_, outer_sep=';', inner_sep=':'):